+++ ts/src/lsp.ts
@@ -415,12 +415,13 @@ const COMPLETION_FUNCTION = 3
 const COMPLETION_KEYWORD = 14
 
 const BUILTIN_FUNCS = [
-  'above', 'acyclic', 'add', 'below', 'close', 'copy', 'deprecate', 'div',
+  'abnf', 'above', 'acyclic', 'add', 'below', 'close', 'copy', 'deprecate',
+  'div',
   'each', 'emit', 'esc',
   'filter', 'greatest',
   'hide', 'inverse', 'join', 'key', 'least', 'length', 'list', 'lower',
   'map', 'match', 'max', 'min', 'mod', 'move', 'mul', 'must', 'neq', 'open',
-  'pack', 'path', 'pick',
+  'pack', 'parse', 'path', 'pick',
   'pref', 're', 'refer', 'rel', 'rem', 'rep', 'split', 'sub', 'sum',
   'super', 'type', 'unique', 'upper', 'usc',
 ]